+++ src/cbp.rs
@@ -19,6 +19,7 @@ fn main() -> anyhow::Result<()> {
         .subcommand(cmd_cbp::check::make_subcommand())
         .subcommand(cmd_cbp::kb::make_subcommand())
         .subcommand(cmd_cbp::tar::make_subcommand())
+        .subcommand(cmd_cbp::upload::make_subcommand())
         .after_help(
             r###"
 Package Manager Features:
@@ -65,6 +66,7 @@ Common Commands:
         Some(("remove", sub_matches)) => cmd_cbp::remove::execute(sub_matches),
         Some(("check", sub_matches)) => cmd_cbp::check::execute(sub_matches),
         Some(("tar", sub_matches)) => cmd_cbp::tar::execute(sub_matches),
+        Some(("upload", sub_matches)) => cmd_cbp::upload::execute(sub_matches),
         _ => unreachable!(),
     }?;