+++ src/cmd_cbp/mod.rs
@@ -8,3 +8,4 @@ pub mod list;
 pub mod local;
 pub mod remove;
 pub mod tar;
+pub mod upload;